pyproject.toml
[tool.poetry]
name = "krofi"
version = "0.1.0"
description = ""
authors = ["abelsouzacosta <abelsouzacostajunior@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pykeepass = "^4.1.0.post1"
python-rofi = "^1.0.1"
pendulum = "^3.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
